// src/app/models/films.ts
export class Films {
    title?: string;
    episode_id?: string;
    opening_crawl?: string;
    director?: string;
    producer?: string;
    release_date?: string;
    characters?: [];
    planets?: [];
    starships?: [];
    vehicles?: [];
    species?: [];
    created?: string;
    edited?: string;
    url?: string;
}